Subindo melhorias 5.2.0

+++ workspace/trabalhoIndividual/montadora/src/main/java/br/com/example/montadora/security/services/EnderecoService.java
@@ -114,7 +114,13 @@ public class EnderecoService {
 //		return enderecoResponseDTO;
 //	}
 	
-	public void deletarEndereco(Integer id) {
-		enderecoRepository.deleteById(id);
+	public boolean deletarEndereco(Integer id) {
+		if(enderecoRepository.existsById(id)) {
+			enderecoRepository.deleteById(id);			
+			return true;
+		} else {
+			return false;
+		}
 	}
+
 }
